split linear models across files
