+++ river/linear_model/__init__.py
@@ -1,7 +1,9 @@
 """Linear models."""
 from .alma import ALMAClassifier
-from .glm import LinearRegression, LogisticRegression, Perceptron
+from .lin_reg import LinearRegression
+from .log_reg import LogisticRegression
 from .pa import PAClassifier, PARegressor
+from .perceptron import Perceptron
 from .softmax import SoftmaxRegression
 
 __all__ = [
